+++ packages/gamut/src/Icon/icons/PandasIcon.tsx
@@ -2,7 +2,7 @@ import React, { SVGProps } from 'react';
 
 import { defaultIconProps } from '../defaultIconProps';
 
-const PandasIcon: React.FC<SVGProps<SVGSVGElement>> = props => {
+export const PandasIcon: React.FC<SVGProps<SVGSVGElement>> = props => {
   return (
     <svg {...props}>
       <title>Panda Icon</title>