+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,137 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.12.0](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.11.0...@codecademy/gamut@2.12.0) (2020-04-01)
+
+
+### Features
+
+* renamed Modal element to Overlay ([#683](https://github.com/RyzacInc/client-modules/issues/683)) ([3bb0ffb](https://github.com/RyzacInc/client-modules/commit/3bb0ffb9d0f078b57012bf72100d026c8c9daf15))
+
+
+
+
+
+# [2.11.0](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.10.0...@codecademy/gamut@2.11.0) (2020-03-31)
+
+
+### Features
+
+* allows 0 width column size for column offset ([#671](https://github.com/RyzacInc/client-modules/issues/671)) ([681a461](https://github.com/RyzacInc/client-modules/commit/681a46148a0d90c0ecec5ad91f1b8ef295a1a982))
+
+
+
+
+
+# [2.10.0](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.9.0...@codecademy/gamut@2.10.0) (2020-03-30)
+
+
+### Features
+
+* use export * on all exported modules ([#658](https://github.com/RyzacInc/client-modules/issues/658)) ([67ba71b](https://github.com/RyzacInc/client-modules/commit/67ba71b54f2d6ff88513a140551d8997a8862c58))
+
+
+
+
+
+# [2.9.0](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.7.1...@codecademy/gamut@2.9.0) (2020-03-24)
+
+
+### Features
+
+* removes link generation from tabs and adds aria attributes ([#661](https://github.com/RyzacInc/client-modules/issues/661)) ([d95e018](https://github.com/RyzacInc/client-modules/commit/d95e018eb941bd2e9e8cf7a72110895ea2cf1f55))
+
+
+
+
+
+# [2.8.0](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.7.1...@codecademy/gamut@2.8.0) (2020-03-24)
+
+
+### Features
+
+* removes link generation from tabs and adds aria attributes ([#661](https://github.com/RyzacInc/client-modules/issues/661)) ([d95e018](https://github.com/RyzacInc/client-modules/commit/d95e018eb941bd2e9e8cf7a72110895ea2cf1f55))
+
+
+
+
+
+## [2.7.1](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.7.0...@codecademy/gamut@2.7.1) (2020-03-10)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+
+
+
+
+# [2.7.0](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.6.1...@codecademy/gamut@2.7.0) (2020-03-05)
+
+
+### Features
+
+* added basic Modal primitive to Gamut ([#641](https://github.com/RyzacInc/client-modules/issues/641)) ([df9fa1f](https://github.com/RyzacInc/client-modules/commit/df9fa1f9c3161b2ef8fcf2e8ff60700c202796cd))
+
+
+
+
+
+## [2.6.1](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.6.0...@codecademy/gamut@2.6.1) (2020-03-04)
+
+
+### Bug Fixes
+
+* fixes responsive prop configs for layout grid ([#648](https://github.com/RyzacInc/client-modules/issues/648)) ([6cd2d23](https://github.com/RyzacInc/client-modules/commit/6cd2d23b8dc07e09f68cf56de5a537f9642c3ad1))
+
+
+
+
+
+# [2.6.0](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.5.0...@codecademy/gamut@2.6.0) (2020-03-04)
+
+
+### Features
+
+* add column offsets to layoutgrid ([#644](https://github.com/RyzacInc/client-modules/issues/644)) ([43e7042](https://github.com/RyzacInc/client-modules/commit/43e7042227271ae337348e4db4fddbeb889c64eb))
+
+
+
+
+
+# [2.5.0](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.4.11...@codecademy/gamut@2.5.0) (2020-03-03)
+
+
+### Features
+
+* added GridForm component ([#605](https://github.com/RyzacInc/client-modules/issues/605)) ([253acad](https://github.com/RyzacInc/client-modules/commit/253acadc5d0775386f48d9290bda77d112c6aa54))
+
+
+
+
+
+## [2.4.11](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.4.10...@codecademy/gamut@2.4.11) (2020-02-24)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+
+
+
+
+## [2.4.10](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.4.9...@codecademy/gamut@2.4.10) (2020-02-24)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+
+
+
+
+## [2.4.9](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.4.8...@codecademy/gamut@2.4.9) (2020-02-24)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+
+
+
+
 ## [2.4.8](https://github.com/RyzacInc/client-modules/compare/@codecademy/gamut@2.4.7...@codecademy/gamut@2.4.8) (2020-02-20)
 
 **Note:** Version bump only for package @codecademy/gamut